[UIMA-6190] Fix bugs found on Mac OSX (using the clang compiler) Explicit casts between the 16-bit characters UChar & XMLCh. Remove unused buggy xmiwriter template. Replace unsigned int with size_t in strtools.hpp Drop -module link flag so Mac creates .dylib rather than .so libraries. Work around Mac's refusal to pass DYLD_LIBRARY_PATH through the environment. Add icu namespace to UnicodeString declarations.

+++ src/framework/uima/strtools.hpp
@@ -1671,8 +1671,8 @@ namespace uima {
     size_t                  pos
   ) {
     const BASIC_STRING_TEMPLATE ENV_VAR = "$ENV(";
-    unsigned int envBegPos = s.find(ENV_VAR, pos);
-    unsigned int envEndPos;
+    size_t envBegPos = s.find(ENV_VAR, pos);
+    size_t envEndPos;
     if (envBegPos != STRING_NPOS) {
       envEndPos = s.find(')', envBegPos+1);
       if (envEndPos != STRING_NPOS) {
